requirements.txt
# Core deep learning
torch>=2.0.0
transformers>=4.30.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0
tensorboard>=2.13.0

# Configuration and utilities
dotenv>=0.9.9
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=1.11.0